@@ -31,7 +31,7 @@
         }
     },
     "require": {
-        "php": "^7.1 || ^8.0",
+        "php": "^7.4 || ^8.0",
         "dealerdirect/phpcodesniffer-composer-installer": "^0.7.0",
         "slevomat/coding-standard": "^7.0.1",
         "squizlabs/php_codesniffer": "^3.6.0"
